refactor(basic): useCart Hook 적용 및 장바구니 로직 분리
