+++ src/basic/hooks/useCart.ts
@@ -1,29 +1,72 @@
 import { useCallback } from 'react';
-import { CartItem, Product } from '../../types';
+import { CartItem, Product, ProductWithUI } from '../../types';
 import { useLocalStorage } from './useLocalStorage';
+import { useProducts } from './useProducts';
 
-export const useCart = () => {
+export const useCart = (
+  addNotification: (
+    message: string,
+    type: 'error' | 'success' | 'warning'
+  ) => void
+) => {
   const [cart, setCart] = useLocalStorage<CartItem[]>('cart', []);
+  const { products } = useProducts();
 
-  const addToCart = useCallback((product: Product) => {
-    setCart(prevCart => {
-      const existingItem = prevCart.find(
-        item => item.product.id === product.id
-      );
-      if (existingItem) {
-        return prevCart.map(item =>
-          item.product.id === product.id
-            ? { ...item, quantity: item.quantity + 1 }
-            : item
-        );
+  const getRemainingStock = useCallback(
+    (product: Product): number => {
+      const cartItem = cart.find((item) => item.product.id === product.id);
+      return product.stock - (cartItem?.quantity || 0);
+    },
+    [cart]
+  );
+
+  const addToCart = useCallback(
+    (product: ProductWithUI) => {
+      const remainingStock = getRemainingStock(product);
+      if (remainingStock <= 0) {
+        addNotification('재고가 부족합니다!', 'error');
+        return;
       }
-      return [...prevCart, { product, quantity: 1 }];
-    });
-  }, []);
 
-  const removeFromCart = useCallback((productId: string) => {
-    setCart(prevCart => prevCart.filter(item => item.product.id !== productId));
-  }, []);
+      setCart((prevCart) => {
+        const existingItem = prevCart.find(
+          (item) => item.product.id === product.id
+        );
+
+        if (existingItem) {
+          const newQuantity = existingItem.quantity + 1;
+
+          if (newQuantity > product.stock) {
+            addNotification(
+              `재고는 ${product.stock}개까지만 있습니다.`,
+              'error'
+            );
+            return prevCart;
+          }
+
+          return prevCart.map((item) =>
+            item.product.id === product.id
+              ? { ...item, quantity: newQuantity }
+              : item
+          );
+        }
+
+        return [...prevCart, { product, quantity: 1 }];
+      });
+
+      addNotification('장바구니에 담았습니다', 'success');
+    },
+    [cart, addNotification, getRemainingStock]
+  );
+
+  const removeFromCart = useCallback(
+    (productId: string) => {
+      setCart((prevCart) =>
+        prevCart.filter((item) => item.product.id !== productId)
+      );
+    },
+    [setCart]
+  );
 
   const updateQuantity = useCallback(
     (productId: string, newQuantity: number) => {
@@ -31,16 +74,26 @@ export const useCart = () => {
         removeFromCart(productId);
         return;
       }
-      setCart(prevCart =>
-        prevCart.map(item =>
+
+      const product = products.find((p) => p.id === productId);
+      if (!product) return;
+
+      const maxStock = product.stock;
+      if (newQuantity > maxStock) {
+        addNotification(`재고는 ${maxStock}개까지만 있습니다.`, 'error');
+        return;
+      }
+
+      setCart((prevCart) =>
+        prevCart.map((item) =>
           item.product.id === productId
             ? { ...item, quantity: newQuantity }
             : item
         )
       );
     },
-    [removeFromCart]
+    [products, removeFromCart, addNotification, setCart]
   );
 
-  return { cart, addToCart, removeFromCart, updateQuantity };
+  return { cart, setCart, addToCart, removeFromCart, updateQuantity, getRemainingStock };
 };
